pyproject.toml
[project]
name = "qupboard-graphql"
version = "0.1.0"
description = "Proof of concept to show how hardware models may be served via GraphQL"
authors = [
    {name = "Jonathan Lucas",email = "jlucas@oqc.tech"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "aiosqlite (>=0.22.1,<0.23.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "fastapi (>=0.131.0,<1.0.0)",
    "pydantic (>=2.4.0,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "strawberry-graphql[cli] (>=0.308.2,<0.309.0)",
    "strawberry-sqlalchemy-mapper (>=0.8.0,<0.9.0)",
    "uvicorn (>=0.41.0,<1.0.0)",
]
[tool.poetry.group.dev.dependencies]
coverage-badge = ">=1.1.2,<2.0.0"
defusedxml = ">=0.7.1,<0.8.0"
genbadge = ">=1.1.3,<2.0.0"
httpx = ">=0.28.1,<0.29.0"
junitparser = ">=4.0.2,<5.0.0"
mkdocs = ">=1.6.1,<2.0.0"
mkdocs-material = ">=9.7.4,<10.0.0"
mkdocstrings = { version = ">=1.0.3,<2.0.0", extras = ["python"] }
pre-commit = ">=4.5.1"
pycobertura = ">=4.1.0,<5.0.0"
pytest = ">=9.0.2"
pytest-asyncio = ">=1.3.0,<2.0.0"
pytest-cov = ">=7.0.0"
pytest-xdist = ">=3.8.0"
ruff = ">=0.15.5"
setuptools = "<=81.0.0"

[project.scripts]
qupboard = "qupboard_graphql.main:main"
[tool.poetry]
packages = [{include = "qupboard_graphql", from = "src"}]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:strawberry_sqlalchemy_mapper",
]

[tool.ruff]
# Allow lines to be as long as 120.
line-length = 120

[tool.coverage.run]
omit = ["src/qupboard_graphql/main.py"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
